refactor: delete the trending movies repo and rely on the movies repo instead
