+++ app/src/main/java/com/baghdad/novix/di/RepositoryModule.kt
@@ -7,7 +7,6 @@ import com.baghdad.domain.repository.FavoriteGenreRepository
 import com.baghdad.domain.repository.MovieRepository
 import com.baghdad.domain.repository.RecentlyViewedRepository
 import com.baghdad.domain.repository.SearchRepository
-import com.baghdad.domain.repository.TrendingMovieRepository
 import com.baghdad.domain.repository.TvShowRepository
 import com.baghdad.repository.ActorRepositoryImpl
 import com.baghdad.repository.ContinueWatchingRepositoryImpl
@@ -16,7 +15,6 @@ import com.baghdad.repository.FavoriteGenreRepositoryImpl
 import com.baghdad.repository.MovieRepositoryImpl
 import com.baghdad.repository.RecentlyViewedRepositoryImpl
 import com.baghdad.repository.SearchRepositoryImpl
-import com.baghdad.repository.TrendingMovieRepositoryImpl
 import com.baghdad.repository.TvShowRepositoryImpl
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
