+++ domain/src/main/java/com/baghdad/domain/repository/MovieRepository.kt
@@ -20,4 +20,5 @@ interface MovieRepository {
     suspend fun getMovieReviews(movieId: Long): List<Review>
     suspend fun getMovieImages(movieId: Long): List<String>
     suspend fun getTopRatedMovies(page: Int): PagedResult<Movie>
+    suspend fun getTrendingMovies(page: Int): PagedResult<Movie>
 }
