--- domain/src/main/java/com/baghdad/domain/repository/TrendingMovieRepository.kt
@@ -1,8 +1,0 @@
-package com.baghdad.domain.repository
-
-import com.baghdad.domain.model.PagedResult
-import com.baghdad.entity.media.Movie
-
-interface TrendingMovieRepository {
-    suspend fun getTrendingMovies(page: Int): PagedResult<Movie>
-}
