+++ domain/src/main/java/com/baghdad/domain/usecase/movie/GetTrendingMoviesUseCase.kt
@@ -1,14 +1,14 @@
 package com.baghdad.domain.usecase.movie
 
 import com.baghdad.domain.model.PagedResult
-import com.baghdad.domain.repository.TrendingMovieRepository
+import com.baghdad.domain.repository.MovieRepository
 import com.baghdad.entity.media.Movie
 
 class GetTrendingMoviesUseCase(
-    private val trendingMovieRepository: TrendingMovieRepository
+    private val movieRepository: MovieRepository
 ) {
     suspend operator fun invoke(page: Int, genreId: Long? = null): PagedResult<Movie> {
-        val pagedResult = trendingMovieRepository.getTrendingMovies(page)
+        val pagedResult = movieRepository.getTrendingMovies(page)
         if (genreId == null) return pagedResult
 
         val filteredData = pagedResult.data.filter { movie ->
